restrict pinned nightly Rust version to only smart contract folder of generated project
