+++ resources/Makefile.in
@@ -1,5 +1,5 @@
 prepare:
-	rustup target add wasm32-unknown-unknown
+	cd contract && rustup target add wasm32-unknown-unknown
 
 build-contract:
 	cd contract && cargo build --release --target wasm32-unknown-unknown
