+++ resources/integration_tests.rs.in
@@ -1,3 +1,7 @@
+fn main() {
+    panic!("Execute \"cargo test\" to test the contract, not \"cargo run\".");
+}
+
 #[cfg(test)]
 mod tests {
     use std::path::PathBuf;
@@ -100,8 +104,4 @@ mod tests {
             actual_error
         );
     }
-}
-
-fn main() {
-    panic!("Execute \"cargo test\" to test the contract, not \"cargo run\".");
 }
